requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django_filter=23.1
psycopg2